consolidate namespaces with identical purpose
